+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 
 
+## 2026-04-23
+- `13:00` -- **Updated** `apps/desktopProbe/src/pages/settings.tsx`
+
 ## 2026-04-19
 - `00:31` -- **Updated** `apps/desktopProbe/src/pages/links.tsx`
 - `00:30` -- **Updated** `apps/desktopProbe/src/pages/links.tsx`